changes co2emssions field of journeyModel
-changes co2emissions field from double to float

+++ app/src/main/java/com/cmpt276/indigo/carbontracker/carbon_tracker_model/JourneyModel.java
@@ -12,7 +12,7 @@ import com.cmpt276.indigo.carbontracker.ListItem;
 public class JourneyModel implements CarbonFootprintComponent{
     private VehicleModel vehicleModel;
     private RouteModel routeModel;
-    private double co2Emission;
+    private float co2Emission;
     private Date creationDate;
 
     public  static List<ListItem> getListItems(){
@@ -24,7 +24,7 @@ public class JourneyModel implements CarbonFootprintComponent{
     public JourneyModel(){
         vehicleModel = new VehicleModel();
         routeModel = new RouteModel();
-        co2Emission = 0;
+        co2Emission = 0.0f;
         creationDate = new Date();
     }
 
@@ -52,11 +52,11 @@ public class JourneyModel implements CarbonFootprintComponent{
         this.vehicleModel = vehicleModel;
     }
 
-    public double getCo2Emission(){
+    public float getCo2Emission(){
         return co2Emission;
     }
 
-    public void setCo2Emission(double co2Emission){
+    public void setCo2Emission(float co2Emission){
         if(co2Emission < 0){
             throw new IllegalArgumentException("CO2 emission cannot be negative.");
         }
